Modelio 3.1.1
git-svn-id: http://forge.modelio.org/svn/modelio3-development-app/branches/modelio3.1@10 17d38812-c249-4409-aa67-f012b2b4a6b9

+++ core.project/src/org/modelio/gproject/model/importer/core/AbstractImporter.java
@@ -234,7 +234,9 @@ public abstract class AbstractImporter {
     @objid ("008e3334-5246-1091-8d81-001ec947cd2a")
     protected void deleteGarbage() {
         for (SmObjectImpl localObjectToDelete : this.result.getDeletions()) {
-            localObjectToDelete.delete();
+            if (localObjectToDelete.isValid()) {
+                localObjectToDelete.delete();
+            }
         }
     }
